pushed header/footer layout down to the individual top-level pages, better react rendering
